解决BUG： 1.同一首歌播放，然后点击切换歌单，但是点击的和播放是同一首，这种情况不切换歌单； 2.获取歌曲url信息的时候状态正常，但是就是没有url的情况下做了预防； 3.Owner类的mid改为Long，防止溢出Int.

+++ app/src/main/java/cn/vce/easylook/feature_music/models/bli/Owner.kt
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName
 class Owner {
     @SerializedName("mid")
     @Expose
-    var mid: Int? = null
+    var mid: Long? = null
 
     @SerializedName("name")
     @Expose
